feat: rename AgentInto to AgentInfo
