+++ frontend/components/MainPageV1/Home/Overview/AgentInfo/AgentActivity/styles.tsx
